Исправление подсчета полупериметра в 2_1_2c

+++ 2_1_2c/Triangle.cpp
@@ -16,7 +16,7 @@ int Triangle::Perimetr()
 
 double Triangle::Square()
 {
-	int p = Perimetr() / 2;
+	double p = Perimetr() / 2.0;
 
 	return sqrt( p * ( p - _a ) * ( p - _b ) * ( p - _c ) );
 }
